Make `/etc/passwd`/`/etc/group` parser streaming
Reduces how much memory is needed to read it and avoids loading more
data than is necessary. The file's normally small enough it's unlikely
to require more than one or two syscalls to read it in full.
