+++ src/common/user_group_tests/mod.rs
@@ -1,3 +1,5 @@
 // User and group tests
-mod parse_etc_group;
-mod parse_etc_passwd;
+mod parse_etc_group_contiguous;
+mod parse_etc_group_split;
+mod parse_etc_passwd_contiguous;
+mod parse_etc_passwd_split;
