+++ src/common/user_group_tests/parse_etc_passwd_contiguous.rs
@@ -1,8 +1,13 @@
-use crate::common::parse_etc_passwd_etc_group as parse;
 use crate::common::IdTable;
 
 fn name(value: &[u8]) -> crate::prelude::IdName {
     crate::prelude::IdName::new(value)
+}
+
+fn parse(value: &[u8]) -> IdTable {
+    let mut parser = crate::common::PasswdGroupParser::new();
+    parser.consume(value);
+    parser.extract()
 }
 
 #[test]
@@ -13,7 +18,7 @@ fn detects_empty_names() {
 #[test]
 fn allows_single_character_names() {
     assert_eq!(
-        parse(b"a:x:123:foo,bar"),
+        parse(b"a:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"a"))])
     );
 }
@@ -21,7 +26,7 @@ fn allows_single_character_names() {
 #[test]
 fn allows_empty_passwords() {
     assert_eq!(
-        parse(b"a::123:foo,bar"),
+        parse(b"a::123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"a"))])
     );
 }
@@ -33,7 +38,10 @@ fn disallows_usernames_not_followed_by_a_colon() {
 
 #[test]
 fn disallows_empty_usernames() {
-    assert_eq!(parse(b":x:123:foo,bar"), IdTable::from_entries(&[]));
+    assert_eq!(
+        parse(b":x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
 }
 
 #[test]
@@ -43,7 +51,10 @@ fn disallows_break_after_password() {
 
 #[test]
 fn disallows_empty_ids() {
-    assert_eq!(parse(b"a:x::foo,bar"), IdTable::from_entries(&[]));
+    assert_eq!(
+        parse(b"a:x::123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
 }
 
 #[test]
@@ -62,23 +73,23 @@ fn allows_break_after_id() {
 #[test]
 fn allows_small_alphanumeric_names() {
     assert_eq!(
-        parse(b"abc123:x:123:foo,bar"),
-        IdTable::from_entries(&[(123, name(b"abc123"))])
+        parse(b"abc123:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[(123, name(b"abc123"))]),
     );
 }
 
 #[test]
 fn allows_small_alphanumeric_names_ending_with_a_dollar_sign() {
     assert_eq!(
-        parse(b"abc123$:x:123:foo,bar"),
-        IdTable::from_entries(&[(123, name(b"abc123$"))])
+        parse(b"abc123$:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[(123, name(b"abc123$"))]),
     );
 }
 
 #[test]
 fn allows_large_alphanumeric_names() {
     assert_eq!(
-        parse(b"abcdefghijklmnopqrstuvwxyz123456:x:123:foo,bar"),
+        parse(b"abcdefghijklmnopqrstuvwxyz123456:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"abcdefghijklmnopqrstuvwxyz123456"))]),
     );
 }
@@ -86,7 +97,7 @@ fn allows_large_alphanumeric_names() {
 #[test]
 fn allows_large_alphanumeric_names_ending_with_a_dollar_sign() {
     assert_eq!(
-        parse(b"abcdefghijklmnopqrstuvwxyz12345$:x:123:foo,bar"),
+        parse(b"abcdefghijklmnopqrstuvwxyz12345$:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"abcdefghijklmnopqrstuvwxyz12345$"))]),
     );
 }
@@ -94,7 +105,7 @@ fn allows_large_alphanumeric_names_ending_with_a_dollar_sign() {
 #[test]
 fn disallows_slightly_too_large_alphanumeric_names() {
     assert_eq!(
-        parse(b"abcdefghijklmnopqrstuvwxyz123456x:x:123:foo,bar"),
+        parse(b"abcdefghijklmnopqrstuvwxyz123456x:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[]),
     );
 }
@@ -102,7 +113,7 @@ fn disallows_slightly_too_large_alphanumeric_names() {
 #[test]
 fn disallows_slightly_too_large_alphanumeric_names_ending_with_a_dollar_sign() {
     assert_eq!(
-        parse(b"abcdefghijklmnopqrstuvwxyz12345x$:x:123:foo,bar"),
+        parse(b"abcdefghijklmnopqrstuvwxyz12345x$:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[]),
     );
 }
@@ -110,7 +121,7 @@ fn disallows_slightly_too_large_alphanumeric_names_ending_with_a_dollar_sign() {
 #[test]
 fn disallows_far_too_large_alphanumeric_names() {
     assert_eq!(
-        parse(b"abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw:x:123:foo,bar"),
+        parse(b"abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[]),
     );
 }
@@ -118,7 +129,9 @@ fn disallows_far_too_large_alphanumeric_names() {
 #[test]
 fn disallows_far_too_large_alphanumeric_names_ending_with_a_dollar_sign() {
     assert_eq!(
-        parse(b"abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw$:x:123:foo,bar"),
+        parse(
+            b"abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw$:x:123:123:::/sbin/nologin",
+        ),
         IdTable::from_entries(&[]),
     );
 }
@@ -126,52 +139,91 @@ fn disallows_far_too_large_alphanumeric_names_ending_with_a_dollar_sign() {
 #[test]
 fn allows_initial_underscores() {
     assert_eq!(
-        parse(b"_user:x:123:foo,bar"),
-        IdTable::from_entries(&[(123, name(b"_user"))])
+        parse(b"_user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[(123, name(b"_user"))]),
     );
 }
 
 #[test]
 fn disallows_initial_dollar_signs() {
-    assert_eq!(parse(b"$user:x:123:foo,bar"), IdTable::from_entries(&[]));
+    assert_eq!(
+        parse(b"$user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
 }
 
 #[test]
 fn disallows_initial_hyphens() {
-    assert_eq!(parse(b"-user:x:123:foo,bar"), IdTable::from_entries(&[]));
+    assert_eq!(
+        parse(b"-user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
 }
 
 #[test]
 fn disallows_initial_numbers() {
-    assert_eq!(parse(b"0user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"1user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"2user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"3user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"4user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"5user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"6user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"7user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"8user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"9user:x:123:foo,bar"), IdTable::from_entries(&[]));
+    assert_eq!(
+        parse(b"0user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"1user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"2user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"3user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"4user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"5user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"6user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"7user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"8user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"9user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
 }
 
 #[test]
 fn disallows_initial_hash_signs() {
-    assert_eq!(parse(b"#user:x:123:foo,bar"), IdTable::from_entries(&[]));
+    assert_eq!(
+        parse(b"#user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
 }
 
 #[test]
 fn disallows_names_with_special_characters_in_the_middle() {
     assert_eq!(
-        parse(b"s!o@m#e%u^s&e*r:x:123:foo,bar"),
-        IdTable::from_entries(&[])
+        parse(b"s!o@m#e%u^s&e*r:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[]),
     );
 }
 
 #[test]
 fn allows_inner_hyphens() {
     assert_eq!(
-        parse(b"some-user:x:123:foo,bar"),
+        parse(b"some-user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"some-user"))]),
     );
 }
@@ -179,15 +231,15 @@ fn allows_inner_hyphens() {
 #[test]
 fn allows_trailing_hyphens() {
     assert_eq!(
-        parse(b"user-:x:123:foo,bar"),
-        IdTable::from_entries(&[(123, name(b"user-"))])
+        parse(b"user-:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[(123, name(b"user-"))]),
     );
 }
 
 #[test]
 fn allows_inner_underscores() {
     assert_eq!(
-        parse(b"some_user:x:123:foo,bar"),
+        parse(b"some_user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"some_user"))]),
     );
 }
@@ -195,8 +247,8 @@ fn allows_inner_underscores() {
 #[test]
 fn allows_trailing_underscores() {
     assert_eq!(
-        parse(b"user_:x:123:foo,bar"),
-        IdTable::from_entries(&[(123, name(b"user_"))])
+        parse(b"user_:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[(123, name(b"user_"))]),
     );
 }
 
@@ -208,16 +260,16 @@ fn detects_empty_names_after_empty_line() {
 #[test]
 fn allows_single_character_names_after_empty_line() {
     assert_eq!(
-        parse(b"\na:x:123:foo,bar"),
-        IdTable::from_entries(&[(123, name(b"a"))])
+        parse(b"\na:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[(123, name(b"a"))]),
     );
 }
 
 #[test]
 fn allows_empty_passwords_after_empty_line() {
     assert_eq!(
-        parse(b"\na::123:foo,bar"),
-        IdTable::from_entries(&[(123, name(b"a"))])
+        parse(b"\na::123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[(123, name(b"a"))]),
     );
 }
 
@@ -228,7 +280,10 @@ fn disallows_usernames_not_followed_by_a_colon_after_empty_line() {
 
 #[test]
 fn disallows_empty_usernames_after_empty_line() {
-    assert_eq!(parse(b"\n:x:123:foo,bar"), IdTable::from_entries(&[]));
+    assert_eq!(
+        parse(b"\n:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
 }
 
 #[test]
@@ -238,7 +293,10 @@ fn disallows_break_after_password_after_empty_line() {
 
 #[test]
 fn disallows_empty_ids_after_empty_line() {
-    assert_eq!(parse(b"\na:x::foo,bar"), IdTable::from_entries(&[]));
+    assert_eq!(
+        parse(b"\na:x::123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
 }
 
 #[test]
@@ -262,15 +320,15 @@ fn allows_break_after_id_end_after_empty_line() {
 #[test]
 fn allows_small_alphanumeric_names_after_empty_line() {
     assert_eq!(
-        parse(b"\nabc123:x:123:foo,bar"),
-        IdTable::from_entries(&[(123, name(b"abc123"))])
+        parse(b"\nabc123:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[(123, name(b"abc123"))]),
     );
 }
 
 #[test]
 fn allows_small_alphanumeric_names_ending_with_a_dollar_sign_after_empty_line() {
     assert_eq!(
-        parse(b"\nabc123$:x:123:foo,bar"),
+        parse(b"\nabc123$:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"abc123$"))]),
     );
 }
@@ -278,7 +336,7 @@ fn allows_small_alphanumeric_names_ending_with_a_dollar_sign_after_empty_line() 
 #[test]
 fn allows_large_alphanumeric_names_after_empty_line() {
     assert_eq!(
-        parse(b"\nabcdefghijklmnopqrstuvwxyz123456:x:123:foo,bar"),
+        parse(b"\nabcdefghijklmnopqrstuvwxyz123456:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"abcdefghijklmnopqrstuvwxyz123456"))]),
     );
 }
@@ -286,7 +344,7 @@ fn allows_large_alphanumeric_names_after_empty_line() {
 #[test]
 fn allows_large_alphanumeric_names_ending_with_a_dollar_sign_after_empty_line() {
     assert_eq!(
-        parse(b"\nabcdefghijklmnopqrstuvwxyz12345$:x:123:foo,bar"),
+        parse(b"\nabcdefghijklmnopqrstuvwxyz12345$:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"abcdefghijklmnopqrstuvwxyz12345$"))]),
     );
 }
@@ -294,7 +352,7 @@ fn allows_large_alphanumeric_names_ending_with_a_dollar_sign_after_empty_line() 
 #[test]
 fn disallows_slightly_too_large_alphanumeric_names_after_empty_line() {
     assert_eq!(
-        parse(b"\nabcdefghijklmnopqrstuvwxyz123456x:x:123:foo,bar"),
+        parse(b"\nabcdefghijklmnopqrstuvwxyz123456x:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[]),
     );
 }
@@ -302,7 +360,7 @@ fn disallows_slightly_too_large_alphanumeric_names_after_empty_line() {
 #[test]
 fn disallows_slightly_too_large_alphanumeric_names_ending_with_a_dollar_sign_after_empty_line() {
     assert_eq!(
-        parse(b"\nabcdefghijklmnopqrstuvwxyz12345x$:x:123:foo,bar"),
+        parse(b"\nabcdefghijklmnopqrstuvwxyz12345x$:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[]),
     );
 }
@@ -310,7 +368,9 @@ fn disallows_slightly_too_large_alphanumeric_names_ending_with_a_dollar_sign_aft
 #[test]
 fn disallows_far_too_large_alphanumeric_names_after_empty_line() {
     assert_eq!(
-        parse(b"\nabcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw:x:123:foo,bar"),
+        parse(
+            b"\nabcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw:x:123:123:::/sbin/nologin",
+        ),
         IdTable::from_entries(&[]),
     );
 }
@@ -318,7 +378,9 @@ fn disallows_far_too_large_alphanumeric_names_after_empty_line() {
 #[test]
 fn disallows_far_too_large_alphanumeric_names_ending_with_a_dollar_sign_after_empty_line() {
     assert_eq!(
-        parse(b"\nabcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw$:x:123:foo,bar"),
+        parse(
+            b"\nabcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw$:x:123:123:::/sbin/nologin",
+        ),
         IdTable::from_entries(&[]),
     );
 }
@@ -326,52 +388,91 @@ fn disallows_far_too_large_alphanumeric_names_ending_with_a_dollar_sign_after_em
 #[test]
 fn allows_initial_underscores_after_empty_line() {
     assert_eq!(
-        parse(b"\n_user:x:123:foo,bar"),
-        IdTable::from_entries(&[(123, name(b"_user"))])
+        parse(b"\n_user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[(123, name(b"_user"))]),
     );
 }
 
 #[test]
 fn disallows_initial_dollar_signs_after_empty_line() {
-    assert_eq!(parse(b"\n$user:x:123:foo,bar"), IdTable::from_entries(&[]));
+    assert_eq!(
+        parse(b"\n$user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
 }
 
 #[test]
 fn disallows_initial_hyphens_after_empty_line() {
-    assert_eq!(parse(b"\n-user:x:123:foo,bar"), IdTable::from_entries(&[]));
+    assert_eq!(
+        parse(b"\n-user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
 }
 
 #[test]
 fn disallows_initial_numbers_after_empty_line() {
-    assert_eq!(parse(b"\n0user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"\n1user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"\n2user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"\n3user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"\n4user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"\n5user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"\n6user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"\n7user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"\n8user:x:123:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"\n9user:x:123:foo,bar"), IdTable::from_entries(&[]));
+    assert_eq!(
+        parse(b"\n0user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"\n1user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"\n2user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"\n3user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"\n4user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"\n5user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"\n6user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"\n7user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"\n8user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
+    assert_eq!(
+        parse(b"\n9user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
 }
 
 #[test]
 fn disallows_initial_hash_signs_after_empty_line() {
-    assert_eq!(parse(b"\n#user:x:123:foo,bar"), IdTable::from_entries(&[]));
+    assert_eq!(
+        parse(b"\n#user:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[])
+    );
 }
 
 #[test]
 fn disallows_names_with_special_characters_in_the_middle_after_empty_line() {
     assert_eq!(
-        parse(b"\ns!o@m#e%u^s&e*r:x:123:foo,bar"),
-        IdTable::from_entries(&[])
+        parse(b"\ns!o@m#e%u^s&e*r:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[]),
     );
 }
 
 #[test]
 fn allows_inner_hyphens_after_empty_line() {
     assert_eq!(
-        parse(b"\nsome-user:x:123:foo,bar"),
+        parse(b"\nsome-user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"some-user"))]),
     );
 }
@@ -379,15 +480,15 @@ fn allows_inner_hyphens_after_empty_line() {
 #[test]
 fn allows_trailing_hyphens_after_empty_line() {
     assert_eq!(
-        parse(b"\nuser-:x:123:foo,bar"),
-        IdTable::from_entries(&[(123, name(b"user-"))])
+        parse(b"\nuser-:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[(123, name(b"user-"))]),
     );
 }
 
 #[test]
 fn allows_inner_underscores_after_empty_line() {
     assert_eq!(
-        parse(b"\nsome_user:x:123:foo,bar"),
+        parse(b"\nsome_user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"some_user"))]),
     );
 }
@@ -395,15 +496,15 @@ fn allows_inner_underscores_after_empty_line() {
 #[test]
 fn allows_trailing_underscores_after_empty_line() {
     assert_eq!(
-        parse(b"\nuser_:x:123:foo,bar"),
-        IdTable::from_entries(&[(123, name(b"user_"))])
+        parse(b"\nuser_:x:123:123:::/sbin/nologin"),
+        IdTable::from_entries(&[(123, name(b"user_"))]),
     );
 }
 
 #[test]
 fn detects_empty_names_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n"),
+        parse(b"user:x:123:123:::/sbin/nologin\n"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
 }
@@ -411,7 +512,7 @@ fn detects_empty_names_after_successful_line() {
 #[test]
 fn allows_single_character_names_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\na:x:456:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\na:x:456:456:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user")), (456, name(b"a"))]),
     );
 }
@@ -419,7 +520,7 @@ fn allows_single_character_names_after_successful_line() {
 #[test]
 fn allows_empty_passwords_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\na::456:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\na::456:456:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user")), (456, name(b"a"))]),
     );
 }
@@ -427,15 +528,15 @@ fn allows_empty_passwords_after_successful_line() {
 #[test]
 fn disallows_usernames_not_followed_by_a_colon_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\na"),
-        IdTable::from_entries(&[(123, name(b"user"))])
+        parse(b"user:x:123:123:::/sbin/nologin\na"),
+        IdTable::from_entries(&[(123, name(b"user"))]),
     );
 }
 
 #[test]
 fn disallows_empty_usernames_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n:x:456:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n:x:456:456:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
 }
@@ -443,7 +544,7 @@ fn disallows_empty_usernames_after_successful_line() {
 #[test]
 fn disallows_break_after_password_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\na:x\n"),
+        parse(b"user:x:123:123:::/sbin/nologin\na:x\n"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
 }
@@ -451,7 +552,7 @@ fn disallows_break_after_password_after_successful_line() {
 #[test]
 fn disallows_empty_ids_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\na:x::foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\na:x::123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
 }
@@ -459,7 +560,7 @@ fn disallows_empty_ids_after_successful_line() {
 #[test]
 fn disallows_break_before_id_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\na:x:\n"),
+        parse(b"user:x:123:123:::/sbin/nologin\na:x:\n"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
 }
@@ -467,7 +568,7 @@ fn disallows_break_before_id_after_successful_line() {
 #[test]
 fn disallows_break_before_id_end_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\na:x:1\n"),
+        parse(b"user:x:123:123:::/sbin/nologin\na:x:1\n"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
 }
@@ -475,7 +576,7 @@ fn disallows_break_before_id_end_after_successful_line() {
 #[test]
 fn allows_break_after_id_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\na:x:1:\n"),
+        parse(b"user:x:123:123:::/sbin/nologin\na:x:1:\n"),
         IdTable::from_entries(&[(123, name(b"user")), (1, name(b"a"))]),
     );
 }
@@ -483,7 +584,7 @@ fn allows_break_after_id_after_successful_line() {
 #[test]
 fn allows_small_alphanumeric_names_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\nabc123:x:456:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\nabc123:x:456:456:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user")), (456, name(b"abc123"))]),
     );
 }
@@ -491,7 +592,7 @@ fn allows_small_alphanumeric_names_after_successful_line() {
 #[test]
 fn allows_small_alphanumeric_names_ending_with_a_dollar_sign_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\nabc123$:x:456:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\nabc123$:x:456:456:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user")), (456, name(b"abc123$"))]),
     );
 }
@@ -500,8 +601,8 @@ fn allows_small_alphanumeric_names_ending_with_a_dollar_sign_after_successful_li
 fn allows_large_alphanumeric_names_after_successful_line() {
     assert_eq!(
         parse(
-            b"user:x:123:foo,bar
-abcdefghijklmnopqrstuvwxyz123456:x:123:foo,bar",
+            b"user:x:123:123:::/sbin/nologin
+abcdefghijklmnopqrstuvwxyz123456:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"abcdefghijklmnopqrstuvwxyz123456"))]),
     );
@@ -511,8 +612,8 @@ abcdefghijklmnopqrstuvwxyz123456:x:123:foo,bar",
 fn allows_large_alphanumeric_names_ending_with_a_dollar_sign_after_successful_line() {
     assert_eq!(
         parse(
-            b"user:x:123:foo,bar
-abcdefghijklmnopqrstuvwxyz12345$:x:123:foo,bar",
+            b"user:x:123:123:::/sbin/nologin
+abcdefghijklmnopqrstuvwxyz12345$:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"abcdefghijklmnopqrstuvwxyz12345$"))]),
     );
@@ -522,8 +623,8 @@ abcdefghijklmnopqrstuvwxyz12345$:x:123:foo,bar",
 fn disallows_slightly_too_large_alphanumeric_names_after_successful_line() {
     assert_eq!(
         parse(
-            b"user:x:123:foo,bar
-abcdefghijklmnopqrstuvwxyz123456x:x:123:foo,bar",
+            b"user:x:123:123:::/sbin/nologin
+abcdefghijklmnopqrstuvwxyz123456x:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
@@ -534,8 +635,8 @@ fn disallows_slightly_too_large_alphanumeric_names_ending_with_a_dollar_sign_aft
 {
     assert_eq!(
         parse(
-            b"user:x:123:foo,bar
-abcdefghijklmnopqrstuvwxyz12345x$:x:123:foo,bar",
+            b"user:x:123:123:::/sbin/nologin
+abcdefghijklmnopqrstuvwxyz12345x$:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
@@ -545,8 +646,8 @@ abcdefghijklmnopqrstuvwxyz12345x$:x:123:foo,bar",
 fn disallows_far_too_large_alphanumeric_names_after_successful_line() {
     assert_eq!(
         parse(
-            b"user:x:123:foo,bar
-abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw:x:123:foo,bar",
+            b"user:x:123:123:::/sbin/nologin
+abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
@@ -556,8 +657,8 @@ abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw:x:123:foo,bar",
 fn disallows_far_too_large_alphanumeric_names_ending_with_a_dollar_sign_after_successful_line() {
     assert_eq!(
         parse(
-            b"user:x:123:foo,bar
-abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw$:x:123:foo,bar",
+            b"user:x:123:123:::/sbin/nologin
+abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw$:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
@@ -566,7 +667,7 @@ abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw$:x:123:foo,bar",
 #[test]
 fn allows_initial_underscores_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n_user:x:456:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n_user:x:456:456:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user")), (456, name(b"_user"))]),
     );
 }
@@ -574,7 +675,7 @@ fn allows_initial_underscores_after_successful_line() {
 #[test]
 fn disallows_initial_dollar_signs_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n$user:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n$user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
 }
@@ -582,7 +683,7 @@ fn disallows_initial_dollar_signs_after_successful_line() {
 #[test]
 fn disallows_initial_hyphens_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n-user:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n-user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
 }
@@ -590,43 +691,43 @@ fn disallows_initial_hyphens_after_successful_line() {
 #[test]
 fn disallows_initial_numbers_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n0user:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n0user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n1user:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n1user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n2user:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n2user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n3user:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n3user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n4user:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n4user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n5user:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n5user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n6user:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n6user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n7user:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n7user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n8user:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n8user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n9user:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n9user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
 }
@@ -634,7 +735,7 @@ fn disallows_initial_numbers_after_successful_line() {
 #[test]
 fn disallows_initial_hash_signs_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\n#user:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\n#user:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
 }
@@ -642,7 +743,7 @@ fn disallows_initial_hash_signs_after_successful_line() {
 #[test]
 fn disallows_names_with_special_characters_in_the_middle_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\ns!o@m#e%u^s&e*r:x:123:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\ns!o@m#e%u^s&e*r:x:123:123:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user"))]),
     );
 }
@@ -650,7 +751,7 @@ fn disallows_names_with_special_characters_in_the_middle_after_successful_line()
 #[test]
 fn allows_inner_hyphens_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\nsome-user:x:456:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\nsome-user:x:456:456:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user")), (456, name(b"some-user"))]),
     );
 }
@@ -658,7 +759,7 @@ fn allows_inner_hyphens_after_successful_line() {
 #[test]
 fn allows_trailing_hyphens_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\nuser-:x:456:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\nuser-:x:456:456:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user")), (456, name(b"user-"))]),
     );
 }
@@ -666,7 +767,7 @@ fn allows_trailing_hyphens_after_successful_line() {
 #[test]
 fn allows_inner_underscores_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\nsome_user:x:456:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\nsome_user:x:456:456:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user")), (456, name(b"some_user"))]),
     );
 }
@@ -674,7 +775,7 @@ fn allows_inner_underscores_after_successful_line() {
 #[test]
 fn allows_trailing_underscores_after_successful_line() {
     assert_eq!(
-        parse(b"user:x:123:foo,bar\nuser_:x:456:foo,bar"),
+        parse(b"user:x:123:123:::/sbin/nologin\nuser_:x:456:456:::/sbin/nologin"),
         IdTable::from_entries(&[(123, name(b"user")), (456, name(b"user_"))]),
     );
 }
@@ -690,7 +791,9 @@ fn detects_empty_names_after_dropped_line() {
 #[test]
 fn allows_single_character_names_after_dropped_line() {
     assert_eq!(
-        parse(b"intentionally invalid line with spaces and !other# n0n$3nse&\na:x:123:foo,bar"),
+        parse(
+            b"intentionally invalid line with spaces and !other# n0n$3nse&\na:x:123:123:::/sbin/nologin"
+        ),
         IdTable::from_entries(&[(123, name(b"a"))]),
     );
 }
@@ -698,7 +801,9 @@ fn allows_single_character_names_after_dropped_line() {
 #[test]
 fn allows_empty_passwords_after_dropped_line() {
     assert_eq!(
-        parse(b"intentionally invalid line with spaces and !other# n0n$3nse&\na::123:foo,bar"),
+        parse(
+            b"intentionally invalid line with spaces and !other# n0n$3nse&\na::123:123:::/sbin/nologin"
+        ),
         IdTable::from_entries(&[(123, name(b"a"))]),
     );
 }
@@ -714,7 +819,9 @@ fn disallows_usernames_not_followed_by_a_colon_after_dropped_line() {
 #[test]
 fn disallows_empty_usernames_after_dropped_line() {
     assert_eq!(
-        parse(b"intentionally invalid line with spaces and !other# n0n$3nse&\n:x:123:foo,bar"),
+        parse(
+            b"intentionally invalid line with spaces and !other# n0n$3nse&\n:x:123:123:::/sbin/nologin"
+        ),
         IdTable::from_entries(&[]),
     );
 }
@@ -732,7 +839,7 @@ fn disallows_empty_ids_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-a:x::foo,bar",
+a:x::123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
@@ -767,7 +874,7 @@ fn allows_small_alphanumeric_names_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-abc123:x:123:foo,bar",
+abc123:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"abc123"))]),
     );
@@ -778,7 +885,7 @@ fn allows_small_alphanumeric_names_ending_with_a_dollar_sign_after_dropped_line(
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-abc123$:x:123:foo,bar",
+abc123$:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"abc123$"))]),
     );
@@ -789,7 +896,7 @@ fn allows_large_alphanumeric_names_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-abcdefghijklmnopqrstuvwxyz123456:x:123:foo,bar",
+abcdefghijklmnopqrstuvwxyz123456:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"abcdefghijklmnopqrstuvwxyz123456"))]),
     );
@@ -800,7 +907,7 @@ fn allows_large_alphanumeric_names_ending_with_a_dollar_sign_after_dropped_line(
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-abcdefghijklmnopqrstuvwxyz12345$:x:123:foo,bar",
+abcdefghijklmnopqrstuvwxyz12345$:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"abcdefghijklmnopqrstuvwxyz12345$"))]),
     );
@@ -811,7 +918,7 @@ fn disallows_slightly_too_large_alphanumeric_names_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-abcdefghijklmnopqrstuvwxyz123456x:x:123:foo,bar",
+abcdefghijklmnopqrstuvwxyz123456x:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
@@ -822,7 +929,7 @@ fn disallows_slightly_too_large_alphanumeric_names_ending_with_a_dollar_sign_aft
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-abcdefghijklmnopqrstuvwxyz12345x$:x:123:foo,bar",
+abcdefghijklmnopqrstuvwxyz12345x$:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
@@ -833,7 +940,7 @@ fn disallows_far_too_large_alphanumeric_names_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw:x:123:foo,bar",
+abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
@@ -844,7 +951,7 @@ fn disallows_far_too_large_alphanumeric_names_ending_with_a_dollar_sign_after_dr
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw$:x:123:foo,bar",
+abcdefghijklmnopqrstuvwxyz123456abcdefghijklmnopqrstuvw$:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
@@ -855,7 +962,7 @@ fn allows_initial_underscores_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-_user:x:123:foo,bar",
+_user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"_user"))]),
     );
@@ -866,7 +973,7 @@ fn disallows_initial_dollar_signs_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-$user:x:123:foo,bar",
+$user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
@@ -877,7 +984,7 @@ fn disallows_initial_hyphens_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
--user:x:123:foo,bar",
+-user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
@@ -888,70 +995,70 @@ fn disallows_initial_numbers_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-0user:x:123:foo,bar",
+0user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-1user:x:123:foo,bar",
+1user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-2user:x:123:foo,bar",
+2user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-3user:x:123:foo,bar",
+3user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-4user:x:123:foo,bar",
+4user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-5user:x:123:foo,bar",
+5user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-6user:x:123:foo,bar",
+6user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-7user:x:123:foo,bar",
+7user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-8user:x:123:foo,bar",
+8user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-9user:x:123:foo,bar",
+9user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
@@ -962,7 +1069,7 @@ fn disallows_initial_hash_signs_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-#user:x:123:foo,bar",
+#user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
@@ -973,7 +1080,7 @@ fn disallows_names_with_special_characters_in_the_middle_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-s!o@m#e%u^s&e*r:x:123:foo,bar",
+s!o@m#e%u^s&e*r:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[]),
     );
@@ -984,7 +1091,7 @@ fn allows_inner_hyphens_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-some-user:x:123:foo,bar",
+some-user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"some-user"))]),
     );
@@ -995,7 +1102,7 @@ fn allows_trailing_hyphens_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-user-:x:123:foo,bar",
+user-:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"user-"))]),
     );
@@ -1006,7 +1113,7 @@ fn allows_inner_underscores_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-some_user:x:123:foo,bar",
+some_user:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"some_user"))]),
     );
@@ -1017,277 +1124,9 @@ fn allows_trailing_underscores_after_dropped_line() {
     assert_eq!(
         parse(
             b"intentionally invalid line with spaces and !other# n0n$3nse&
-user_:x:123:foo,bar",
+user_:x:123:123:::/sbin/nologin",
         ),
         IdTable::from_entries(&[(123, name(b"user_"))]),
-    );
-}
-
-#[test]
-fn id_parser_works_for_unpadded_powers_of_10() {
-    assert_eq!(
-        parse(b"user:x:10:foo,bar"),
-        IdTable::from_entries(&[(10, name(b"user"))])
-    );
-    assert_eq!(
-        parse(b"user:x:100:foo,bar"),
-        IdTable::from_entries(&[(100, name(b"user"))])
-    );
-    assert_eq!(
-        parse(b"user:x:1000:foo,bar"),
-        IdTable::from_entries(&[(1000, name(b"user"))])
-    );
-    assert_eq!(
-        parse(b"user:x:10000:foo,bar"),
-        IdTable::from_entries(&[(10000, name(b"user"))])
-    );
-    assert_eq!(
-        parse(b"user:x:100000:foo,bar"),
-        IdTable::from_entries(&[(100000, name(b"user"))])
-    );
-    assert_eq!(
-        parse(b"user:x:1000000:foo,bar"),
-        IdTable::from_entries(&[(1000000, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:10000000:foo,bar"),
-        IdTable::from_entries(&[(10000000, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:100000000:foo,bar"),
-        IdTable::from_entries(&[(100000000, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:1000000000:foo,bar"),
-        IdTable::from_entries(&[(1000000000, name(b"user"))]),
-    );
-}
-
-#[test]
-fn id_parser_works_for_padded_powers_of_10() {
-    assert_eq!(
-        parse(b"user:x:0000000001:foo,bar"),
-        IdTable::from_entries(&[(1, name(b"user"))])
-    );
-    assert_eq!(
-        parse(b"user:x:0000000010:foo,bar"),
-        IdTable::from_entries(&[(10, name(b"user"))])
-    );
-    assert_eq!(
-        parse(b"user:x:0000000100:foo,bar"),
-        IdTable::from_entries(&[(100, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:0000001000:foo,bar"),
-        IdTable::from_entries(&[(1000, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:0000010000:foo,bar"),
-        IdTable::from_entries(&[(10000, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:0000100000:foo,bar"),
-        IdTable::from_entries(&[(100000, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:0001000000:foo,bar"),
-        IdTable::from_entries(&[(1000000, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:0010000000:foo,bar"),
-        IdTable::from_entries(&[(10000000, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:0100000000:foo,bar"),
-        IdTable::from_entries(&[(100000000, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:1000000000:foo,bar"),
-        IdTable::from_entries(&[(1000000000, name(b"user"))]),
-    );
-}
-
-#[test]
-fn id_parser_works_for_unpadded_powers_of_10_plus_1() {
-    assert_eq!(
-        parse(b"user:x:11:foo,bar"),
-        IdTable::from_entries(&[(11, name(b"user"))])
-    );
-    assert_eq!(
-        parse(b"user:x:101:foo,bar"),
-        IdTable::from_entries(&[(101, name(b"user"))])
-    );
-    assert_eq!(
-        parse(b"user:x:1001:foo,bar"),
-        IdTable::from_entries(&[(1001, name(b"user"))])
-    );
-    assert_eq!(
-        parse(b"user:x:10001:foo,bar"),
-        IdTable::from_entries(&[(10001, name(b"user"))])
-    );
-    assert_eq!(
-        parse(b"user:x:100001:foo,bar"),
-        IdTable::from_entries(&[(100001, name(b"user"))])
-    );
-    assert_eq!(
-        parse(b"user:x:1000001:foo,bar"),
-        IdTable::from_entries(&[(1000001, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:10000001:foo,bar"),
-        IdTable::from_entries(&[(10000001, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:100000001:foo,bar"),
-        IdTable::from_entries(&[(100000001, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:1000000001:foo,bar"),
-        IdTable::from_entries(&[(1000000001, name(b"user"))]),
-    );
-}
-
-#[test]
-fn id_parser_works_for_padded_powers_of_10_plus_1() {
-    assert_eq!(
-        parse(b"user:x:0000000011:foo,bar"),
-        IdTable::from_entries(&[(11, name(b"user"))])
-    );
-    assert_eq!(
-        parse(b"user:x:0000000101:foo,bar"),
-        IdTable::from_entries(&[(101, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:0000001001:foo,bar"),
-        IdTable::from_entries(&[(1001, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:0000010001:foo,bar"),
-        IdTable::from_entries(&[(10001, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:0000100001:foo,bar"),
-        IdTable::from_entries(&[(100001, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:0001000001:foo,bar"),
-        IdTable::from_entries(&[(1000001, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:0010000001:foo,bar"),
-        IdTable::from_entries(&[(10000001, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:0100000001:foo,bar"),
-        IdTable::from_entries(&[(100000001, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:1000000001:foo,bar"),
-        IdTable::from_entries(&[(1000000001, name(b"user"))]),
-    );
-}
-
-#[test]
-fn id_parser_works_near_u32_representation_limit() {
-    assert_eq!(
-        parse(b"user:x:4294967290:foo,bar"),
-        IdTable::from_entries(&[(4294967290, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:4294967291:foo,bar"),
-        IdTable::from_entries(&[(4294967291, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:4294967292:foo,bar"),
-        IdTable::from_entries(&[(4294967292, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:4294967293:foo,bar"),
-        IdTable::from_entries(&[(4294967293, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:4294967294:foo,bar"),
-        IdTable::from_entries(&[(4294967294, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:4294967295:foo,bar"),
-        IdTable::from_entries(&[(4294967295, name(b"user"))]),
-    );
-    assert_eq!(
-        parse(b"user:x:4294967296:foo,bar"),
-        IdTable::from_entries(&[])
-    );
-    assert_eq!(
-        parse(b"user:x:4294967297:foo,bar"),
-        IdTable::from_entries(&[])
-    );
-    assert_eq!(
-        parse(b"user:x:4294967298:foo,bar"),
-        IdTable::from_entries(&[])
-    );
-    assert_eq!(
-        parse(b"user:x:4294967299:foo,bar"),
-        IdTable::from_entries(&[])
-    );
-}
-
-#[test]
-fn id_parser_rejects_numbers_suffixed_with_hyphens() {
-    assert_eq!(parse(b"user:x:0-:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"user:x:10-:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"user:x:100-:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"user:x:1000-:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"user:x:10000-:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"user:x:100000-:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(
-        parse(b"user:x:1000000-:foo,bar"),
-        IdTable::from_entries(&[])
-    );
-    assert_eq!(
-        parse(b"user:x:10000000-:foo,bar"),
-        IdTable::from_entries(&[])
-    );
-    assert_eq!(
-        parse(b"user:x:100000000-:foo,bar"),
-        IdTable::from_entries(&[])
-    );
-    assert_eq!(
-        parse(b"user:x:429496729-:foo,bar"),
-        IdTable::from_entries(&[])
-    );
-    assert_eq!(
-        parse(b"user:x:4294967295-:foo,bar"),
-        IdTable::from_entries(&[])
-    );
-}
-
-#[test]
-fn id_parser_rejects_numbers_prefixed_with_hyphens() {
-    assert_eq!(parse(b"user:x:-0:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"user:x:-10:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"user:x:-100:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"user:x:-1000:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"user:x:-10000:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(parse(b"user:x:-100000:foo,bar"), IdTable::from_entries(&[]));
-    assert_eq!(
-        parse(b"user:x:-1000000:foo,bar"),
-        IdTable::from_entries(&[])
-    );
-    assert_eq!(
-        parse(b"user:x:-10000000:foo,bar"),
-        IdTable::from_entries(&[])
-    );
-    assert_eq!(
-        parse(b"user:x:-100000000:foo,bar"),
-        IdTable::from_entries(&[])
-    );
-    assert_eq!(
-        parse(b"user:x:-429496729:foo,bar"),
-        IdTable::from_entries(&[])
-    );
-    assert_eq!(
-        parse(b"user:x:-4294967295:foo,bar"),
-        IdTable::from_entries(&[])
     );
 }
