+++ src/test_utils/mod.rs
@@ -3,6 +3,7 @@
 mod assert;
 mod errors;
 mod get_user_group_table;
+mod misc;
 mod spy;
 mod thread_checkpoint;
 mod time;
@@ -11,6 +12,7 @@ mod with_attempts;
 pub use assert::*;
 pub use errors::*;
 pub use get_user_group_table::*;
+pub use misc::*;
 pub use spy::*;
 pub use thread_checkpoint::*;
 pub use time::*;
